my project

+++ src/main/java/com/javaee/project/project3/controller/ProductCategoryController.java
@@ -1,11 +1,18 @@
 package com.javaee.project.project3.controller;
 
 import com.javaee.project.project3.form.FieldName;
+import com.javaee.project.project3.model.User;
 import com.javaee.project.project3.service.ProductCategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Controller
 @RequestMapping(value = "categorys")
@@ -14,39 +21,43 @@ public class ProductCategoryController {
     private ProductCategoryService categoryService;
 
     @GetMapping
-    public String getAll(Model model){
-        FieldName category=new FieldName();
-        model.addAttribute("categories",categoryService.getAllCategories());
-        model.addAttribute("category",category);
+    public String getAll(Model model) {
+        FieldName category = new FieldName();
+        model.addAttribute("categories", categoryService.getAllCategories());
+        model.addAttribute("category", category);
         return "category/categories";
     }
+
     @PostMapping("create")
-    public String create(@ModelAttribute FieldName category){
+    public String create(@ModelAttribute FieldName category) {
         categoryService.create(category.getField());
         return "redirect:";
     }
 
     @GetMapping(value = "delete/{id}")
-    public String deleteCategory(@PathVariable(name = "id") Long id, Model model){
-        categoryService.deleteById(id);
+    public String deleteCategory(@PathVariable(name = "id") Long id, Model model) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User sds = ((User) principal);
+        if (sds.getRoles().stream().anyMatch(o->o.getRole().contains("ROLE_ADMIN")))
+            categoryService.deleteById(id);
         return getAll(model);
     }
 
     @GetMapping(value = "edit/{id}")
-    public String editCategory(@PathVariable(name = "id")Long id, Model model){
-        FieldName category=new FieldName();
+    public String editCategory(@PathVariable(name = "id") Long id, Model model) {
+        FieldName category = new FieldName();
         model.addAttribute("categoryId", id);
-        model.addAttribute("category",category);
+        model.addAttribute("category", category);
         return "category/editCategory";
     }
 
     @PostMapping(value = "edit/{id}")
-    public String saveCategory(@PathVariable(name = "id") Long id, @ModelAttribute FieldName category, Model model){
+    public String saveCategory(@PathVariable(name = "id") Long id, @ModelAttribute FieldName category, Model model) {
 
-        categoryService.updateCategory(id,category.getField());
+        categoryService.updateCategory(id, category.getField());
         return getAll(model);
     }
-
 
 
 }
